% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calendar_freebusy.R
\name{calendar_freebusy}
\alias{calendar_freebusy}
\title{return availability for a list of calendars}
\usage{
calendar_freebusy(
  ids,
  time_min = NULL,
  time_max = NULL,
  token = calendar_token()
)
}
\arguments{
\item{ids}{a character vector of calendars or groups to query}

\item{time_min}{min date}

\item{time_max}{max date}

\item{token}{token obtained from \code{\link[=calendar_token]{calendar_token()}}}
}
\value{
a ti
}
\description{
return availability for a list of calendars
}
