% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calendar_list.R
\name{calendar_list}
\alias{calendar_list}
\title{Get list of calendars from google.}
\usage{
calendar_list(token = calendar_token())
}
\arguments{
\item{token}{token obtained from \code{\link[=calendar_token]{calendar_token()}}}
}
\value{
Tibble with columns for \code{id}, \code{summary} and \code{description}.
}
\description{
Get list of calendars from google.
}
\seealso{
Other calendar functions: 
\code{\link{calendar_events}()}
}
\concept{calendar functions}
