% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalendr_auth.R
\name{calendar_token}
\alias{calendar_token}
\title{Produce configured token}
\usage{
calendar_token()
}
\value{
A \code{request} object (an S3 class provided by \link[httr:httr-package]{httr}).
}
\description{
For internal use or for those programming around the Calendar API.
Returns a token pre-processed with \code{\link[httr:config]{httr::config()}}. Most users
do not need to handle tokens "by hand" or, even if they need some
control, \code{\link[=calendar_auth]{calendar_auth()}} is what they need. If there is no current
token, \code{\link[=calendar_auth]{calendar_auth()}} is called to either load from cache or
initiate OAuth2.0 flow.
If auth has been deactivated via \code{\link[=calendar_deauth]{calendar_deauth()}}, \code{calendar_token()}
returns \code{NULL}.
}
\examples{
\dontrun{
req <- request_generate(
  "drive.files.get",
  list(fileId = "abc"),
  token = calendar_token()
)
req
}
}
\seealso{
Other low-level API functions: 
\code{\link{request_generate}()}
}
\concept{low-level API functions}
