% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_event.R
\name{create_event}
\alias{create_event}
\title{Create a Google calendar event}
\usage{
create_event(
  id,
  title,
  description,
  attendees,
  start_time,
  duration,
  timezone = Sys.timezone()
)
}
\arguments{
\item{id}{calendar id, obtained from \code{\link[=calendar_list]{calendar_list()}}}

\item{title}{Title of the event. Optional.}

\item{description}{description of the event. This should be a character vector. Optional.}

\item{attendees}{character vector of email addresses of event attendees. Optional.}

\item{start_time}{the event start time formatted "yyyy-mm-dd hh:mm:ss". Hours, minutes, seconds, optional.}

\item{duration}{duration of calendar event in minutes. Optional.}

\item{timezone}{See \code{?timezones}.}
}
\description{
This function allows you to create a calendar event.
}
