% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enlist}
\alias{enlist}
\title{convert character vector of ids to named list}
\usage{
enlist(x)
}
\arguments{
\item{x}{a character vector of calendar ids}
}
\value{
a list character vectors, each with name "id"
}
\description{
convert character vector of ids to named list
}
