% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmt_time}
\alias{fmt_time}
\title{format time for calendar API}
\usage{
fmt_time(date, zone = "UTC")
}
\arguments{
\item{time}{a POSIXct or POSIXlt object}

\item{tz}{an IANA timezone designation}
}
\value{
an RFC 3339 formatted character vector
}
\description{
format time for calendar API
}
