% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_google_token}
\alias{get_google_token}
\title{Defunct.}
\usage{
get_google_token(google_app)
}
\arguments{
\item{google_app}{No longer used}
}
\description{
Please use \code{\link[=gcalendr_auth]{gcalendr_auth()}} and \code{\link[=gcalendr_token]{gcalendr_token()}} instead.
}
