% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalendr-package.r
\docType{package}
\name{gcalendr}
\alias{gcalendr}
\alias{gcalendr-package}
\title{gcalendr}
\description{
You can read Google calendar data using the  Google Calendar API
}
\details{
Access your calendar:
\itemize{
\item \code{\link[=get_gcal_list]{get_gcal_list()}}
\item \code{\link[=get_gcal_events]{get_gcal_events()}}
}

Authenticate
\itemize{
\item \code{\link[=gcalendr_auth]{gcalendr_auth()}}
\item \code{\link[=gcalendr_token]{gcalendr_token()}}
}
}
