% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalendr_auth.R
\name{gcalendr_auth}
\alias{gcalendr_auth}
\title{Authorize gcalendr}
\usage{
gcalendr_auth(email = NULL, path = NULL,
  scopes = "https://www.googleapis.com/auth/drive",
  cache = gargle::gargle_oauth_cache(),
  use_oob = gargle::gargle_oob_default(), token = NULL)
}
\description{
Authorize gcalendr to view and manage your calendar events. By default,
you are directed to a web browser, asked to sign in to your Google
account, and to grant gcalendr permission to operate on your behalf
with Google Calendar. By default, these user credentials are cached in a
folder below your home directory, \code{~/.R/gargle/gargle-oauth}, from where
they can be automatically refreshed, as necessary. Storage at the user
level means the same token can be used across multiple projects and
tokens are less likely to be synced to the cloud by accident.
}
\details{
Most users, most of the time, do not need to call \code{gcalendr_auth()}
explicitly -- it is triggered by the first action that requires
authorization. Even when called, the default arguments often suffice.
However, when necessary, this function allows the user to explicitly:
\itemize{
\item Declare which Google identity to use, via an email address. If there
are multiple cached tokens, this can clarify which one to use. It can
also force gcalendr to switch from one identity to another. If
there's no cached token for the email, this triggers a return to the
browser to choose the identity and give consent.
\item Use a service account token.
\item Specify non-default behavior re: token caching and out-of-bound
authentication.
}

For even deeper control over auth, use \code{\link[=gcalendr_auth_config]{gcalendr_auth_config()}} to use
your own OAuth app or API key.
}
\examples{
\dontrun{
## load/refresh existing credentials, if available
## otherwise, go to browser for authentication and authorization
gcalendr_auth()

## see user associated with current token
gcalendr_user()

## force use of a token associated with a specific email
gcalendr_auth(email = "jenny@example.com")
gcalendr_user()

## force a menu where you can choose from existing tokens or
## choose to get a new one
gcalendr_auth(email = NA)

## use a 'read only' scope, so it's impossible to edit or delete files
gcalendr_auth(
  scopes = "https://www.googleapis.com/auth/drive.readonly"
)

## use a service account token
gcalendr_auth(path = "foofy-83ee9e7c9c48.json")
}
}
\seealso{
Other auth functions: \code{\link{gcalendr_auth_config}},
  \code{\link{gcalendr_deauth}}
}
\concept{auth functions}
