% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalendr_auth.R
\name{gcalendr_deauth}
\alias{gcalendr_deauth}
\title{Suspend authorization}
\usage{
gcalendr_deauth()
}
\description{
Put gcalendr into a de-authorized state. Instead of sending a token,
gcalendr will send its API key. This can be used to access public
files for which no Google sign-in is required. This is handy for using
gcalendr in a non-interactive setting to make requests that do not
require a token. It will prevent the attempt to obtain a token
interactively in the browser. A built-in API key is used by default or
the user can configure their own via \code{\link[=gcalendr_auth_config]{gcalendr_auth_config()}}.
}
\examples{
\dontrun{
gcalendr_deauth()
gcalendr_user()
public_file <-
  gcalendr_get(as_id("1Hj-k7NpPSyeOR3R7j4KuWnru6kZaqqOAE8_db5gowIM"))
gcalendr_download(public_file)
}
}
\seealso{
Other auth functions: \code{\link{gcalendr_auth_config}},
  \code{\link{gcalendr_auth}}
}
\concept{auth functions}
