% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalendr_endpoints.R
\name{gcalendr_endpoints}
\alias{gcalendr_endpoints}
\title{List Calendar endpoints}
\usage{
gcalendr_endpoints(i = NULL)
}
\arguments{
\item{i}{The name(s) or integer index(ices) of the endpoints to return.
Optional. By default, the entire list is returned.}
}
\value{
A list containing some or all of the subset of the calendar API v3
endpoints that are used internally by \code{gcalendr}.
}
\description{
Returns a list of selected calendar API v3 endpoints, as stored inside the
googlecalendar package. The names of this list (or the \code{id} sub-elements) are
the nicknames that can be used to specify an endpoint in
\code{\link[=request_generate]{request_generate()}}. For each endpoint, we store its nickname or \code{id}, the
associated HTTP verb, the \code{path}, and details about the parameters. This list
is derived programmatically from the
\href{https://www.googleapis.com/discovery/v1/apis/calendar/v3/rest}{calendar API v3 Discovery Document}.
}
\examples{
str(gcalendr_endpoints(), max.level = 2)
gcalendr_endpoints("calendar.events.list")
gcalendr_endpoints(4)
}
