% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalendr_auth.R
\name{gcalendr_has_token}
\alias{gcalendr_has_token}
\title{Is there a token on hand?}
\usage{
gcalendr_has_token()
}
\value{
Logical.
}
\description{
Reports whether gcalendr has stored a token, ready for use in downstream
requests. Exists mostly for protecting examples that won't work in the
absence of a token.
}
\examples{
gcalendr_has_token()
}
