% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalendr_auth.R
\name{gcalendr_token}
\alias{gcalendr_token}
\title{Produce configured token}
\usage{
gcalendr_token()
}
\value{
A \code{request} object (an S3 class provided by \link[httr:httr]{httr}).
}
\description{
For internal use or for those programming around the Calendar API.
Returns a token pre-processed with \code{\link[httr:config]{httr::config()}}. Most users
do not need to handle tokens "by hand" or, even if they need some
control, \code{\link[=gcalendr_auth]{gcalendr_auth()}} is what they need. If there is no current
token, \code{\link[=gcalendr_auth]{gcalendr_auth()}} is called to either load from cache or
initiate OAuth2.0 flow. If auth has been deactivated via
\code{\link[=gcalendr_auth_config]{gcalendr_auth_config()}}, \code{gcalendr_token()} returns \code{NULL}.
}
\examples{
\dontrun{
req <- request_generate(
  "drive.files.get",
  list(fileId = "abc"),
  token = gcalendr_token()
)
req
}
}
\seealso{
Other low-level API functions: \code{\link{generate_request}}
}
\concept{low-level API functions}
