% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calendar_events.R
\name{get_gcal_events}
\alias{get_gcal_events}
\title{Read list of google calendar events.}
\usage{
get_gcal_events(id, token = gcalendr_token(), days_in_past = 90,
  days_in_future = 90, now = Sys.time(), max_results = 250)
}
\arguments{
\item{id}{calendar id, obtained from \code{\link[=get_gcal_list]{get_gcal_list()}}}

\item{token}{token obtained from \link{get_google_token}}

\item{days_in_past}{Restrict results to date range, number of days in past}

\item{days_in_future}{Restrict results to date range, number of days into future}

\item{now}{Reference time stamp, defaulting to \code{\link[=Sys.time]{Sys.time()}}}

\item{max_results}{Maximum number of results retrieved per query}
}
\description{
Read list of google calendar events.
}
\references{
https://developers.google.com/google-apps/calendar/v3/reference/events/list
}
\seealso{
Other gcal functions: \code{\link{get_gcal_list}}
}
\concept{gcal functions}
