% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calendar_list.R
\name{get_gcal_list}
\alias{get_gcal_list}
\title{Get list of calendars from google.}
\usage{
get_gcal_list(token = gcalendr_token())
}
\arguments{
\item{token}{token obtained from \code{\link[=gcalendr_token]{gcalendr_token()}}}
}
\value{
Tibble with columns for \code{id}, \code{summary} and \code{description}.
}
\description{
Get list of calendars from google.
}
\seealso{
Other calendar functions: \code{\link{get_gcal_events}}
}
\concept{calendar functions}
