% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_google_calendar.R
\name{get_google_token}
\alias{get_google_token}
\title{Get google token for authentication.}
\usage{
get_google_token()
}
\description{
Get google token for authentication.
}
\seealso{
Other gcal functions: \code{\link{get_gcal_events}},
  \code{\link{get_gcal_list}}
}
