% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalendr_auth.R
\name{calendar_has_token}
\alias{calendar_has_token}
\title{Is there a token on hand?}
\usage{
calendar_has_token()
}
\value{
Logical.

Logical.
}
\description{
Reports whether gcalendr has stored a token, ready for use in downstream
requests.
}
\examples{
calendar_has_token()
}
