% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{get_google_token}
\alias{get_google_token}
\alias{get_gcal_list}
\alias{get_gcal_events}
\title{Functions that are no longer used.}
\usage{
get_google_token(google_app)

get_gcal_list(token = calendar_token())

get_gcal_events(
  id,
  token = calendar_token(),
  days_in_past = 90,
  days_in_future = 90,
  now = Sys.Date(),
  max_results = 250
)
}
\arguments{
\item{google_app}{No longer used}

\item{token}{token obtained from \code{\link[=calendar_token]{calendar_token()}}}

\item{id}{calendar id, obtained from \code{\link[=calendar_list]{calendar_list()}}}

\item{days_in_past}{Restrict results to date range, number of days in past}

\item{days_in_future}{Restrict results to date range, number of days into
future}

\item{now}{Reference time stamp, defaulting to \code{\link[=Sys.time]{Sys.time()}}}

\item{max_results}{Used internally to specify the maximum number of results
retrieved per "page" of the query.}
}
\description{
These functions were available in a very early version of the package.
Please use these replacement functions instead:
\itemize{
\item \code{\link[=calendar_auth]{calendar_auth()}} and \code{\link[=calendar_token]{calendar_token()}}
\item \code{\link[=calendar_list]{calendar_list()}}
\item \code{\link[=calendar_events]{calendar_events()}}
}
}
