% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalendr-package.r
\docType{package}
\name{gcalendr}
\alias{gcalendr}
\alias{gcalendr-package}
\title{gcalendr}
\description{
You can read Google calendar data using the  Google Calendar API
}
\details{
Access your calendar:
\itemize{
\item \code{\link[=calendar_list]{calendar_list()}}
\item \code{\link[=calendar_events]{calendar_events()}}
}

Authenticate
\itemize{
\item \code{\link[=calendar_auth]{calendar_auth()}}
\item \code{\link[=calendar_token]{calendar_token()}}
}
}
