% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcalendr-package.r
\docType{package}
\name{gcalendr}
\alias{gcalendr}
\alias{gcalendr-package}
\title{gcalendr}
\description{
Important functions:
}
\details{
\itemize{
\item \code{\link[=get_google_token]{get_google_token()}}
\item \code{\link[=get_gcal_list]{get_gcal_list()}}
\item \code{\link[=get_gcal_events]{get_gcal_events()}}
}
}
