% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_google_calendar.R
\name{get_gcal_events}
\alias{get_gcal_events}
\title{Read list of google calendar events.}
\usage{
get_gcal_events(id, google_token, max_results = 250, days_in_past = 90,
  days_in_future = 90)
}
\arguments{
\item{id}{calendar id, obtained from \code{\link[=get_gcal_list]{get_gcal_list()}}}

\item{google_token}{token obtained from \link{get_google_token}}

\item{max_results}{Maximum number of results retrieved per query}

\item{days_in_past}{Restrict results to date range, number of days in past}

\item{days_in_future}{Restrict results to date range, number of days into future}
}
\description{
Read list of google calendar events.
}
\references{
https://developers.google.com/google-apps/calendar/v3/reference/events/list
}
\seealso{
Other gcal functions: \code{\link{get_gcal_list}},
  \code{\link{get_google_token}}
}
\concept{gcal functions}
