% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_google_calendar.R
\name{get_gcal_list}
\alias{get_gcal_list}
\title{Get list of calendars from google.}
\usage{
get_gcal_list(google_token)
}
\arguments{
\item{google_token}{token obtained from \link{get_google_token}}
}
\value{
character vector with calendar ids
}
\description{
Get list of calendars from google.
}
\seealso{
Other gcal functions: \code{\link{get_gcal_events}},
  \code{\link{get_google_token}}
}
\concept{gcal functions}
