% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calendar_list.R
\name{get_gcal_list}
\alias{get_gcal_list}
\title{Get list of calendars from google.}
\usage{
get_gcal_list(token = gcalendr_token())
}
\arguments{
\item{token}{token obtained from \link{get_google_token}}
}
\value{
character vector with calendar ids
}
\description{
Get list of calendars from google.
}
\seealso{
Other gcal functions: \code{\link{get_gcal_events}}
}
\concept{gcal functions}
