% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_google_token}
\alias{get_google_token}
\title{Get google token for authentication.}
\usage{
get_google_token(google_app)
}
\arguments{
\item{google_app}{A valid oauth app, e.g. as provided by \link[httr:oauth_app]{httr::oauth_app}.  Defaults to an app that will ask for permission to view your google calendars.  You can use the built-in app, or create your own.}
}
\description{
Get google token for authentication.
}
\seealso{
Other gcal functions: \code{\link{get_gcal_events}},
  \code{\link{get_gcal_list}}
}
\concept{gcal functions}
